% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_mtp}
\alias{c_mtp}
\title{Multiple testing following the graph}
\usage{
c_mtp(p_values, alpha, G, log = FALSE)
}
\arguments{
\item{p_values}{vector of p-values for the elementary hypothesis}

\item{alpha}{vector of the original alphas}

\item{G}{transition matrix G}

\item{log}{TRUE: print log at each step; FALSE: silent}
}
\value{
Hypothesis rejection status indicator vector
}
\description{
Multiple testing following the graph
}
