% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_optim.R
\name{om_rejection_optim}
\alias{om_rejection_optim}
\title{Optimize weights and G}
\usage{
om_rejection_optim(weights, mat_g, ..., tot = 1, par_optim = list())
}
\arguments{
\item{weights}{vector of the original weights}

\item{mat_g}{transition matrix G}

\item{tot}{restrict on the sum of parameters}

\item{par_optim}{options for constrOptim function}
}
\description{
Optimize weights and G
}
