% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_simu.R
\name{om_simu}
\alias{om_simu}
\title{Simulate endpoint}
\usage{
om_simu(n, s_mu, s_sd, s_cor, seed = NULL)
}
\arguments{
\item{n}{number of replications}

\item{s_mu}{simulation mean}

\item{s_sd}{simulation SD}

\item{s_cor}{correlation}
}
\value{
Matrix of simulated test statistics
}
\description{
Simulate endpoint
}
