% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_optim.R
\name{om_theta_fill}
\alias{om_theta_fill}
\title{Fill parameters to the G and weights}
\usage{
om_theta_fill(theta = NULL, weights, mat_g, tot = 1)
}
\arguments{
\item{theta}{parameters in weights and G}

\item{weights}{vector of the original weights}

\item{mat_g}{transition matrix G}
}
\description{
Fill parameters to the G and weights
}
