% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_optim.R
\name{om_theta_rejection}
\alias{om_theta_rejection}
\title{Fill parameter then calculate rejection}
\usage{
om_theta_rejection(theta, weights, mat_g, ...)
}
\arguments{
\item{theta}{parameters in weights and G}

\item{weights}{vector of the original weights}

\item{mat_g}{transition matrix G}
}
\description{
Fill parameter then calculate rejection
}
